[#10] Add flags to configure secure connection client certificate
- unified implementations of http test servers
- minor refactoring and improvements
