+++ internal/storage/download_test.go
@@ -12,25 +12,11 @@
 package storage
 
 import (
-	"context"
-	"fmt"
-	"net/http"
-	"net/url"
 	"os"
 	"path/filepath"
 	"runtime"
-	"strconv"
 	"strings"
-	"sync"
 	"testing"
-	"time"
-)
-
-var (
-	// testOnceRange is used to add HTTP alias only once for the range server.
-	testOnceRange sync.Once
-	// testOnceSimple is used to add HTTP alias only once for the simple server.
-	testOnceSimple sync.Once
 )
 
 const (
@@ -41,13 +27,6 @@ const (
 	extCert     = "testdata/ext_cert.pem"
 	extKey      = "testdata/ext_key.pem"
 )
-
-type web struct {
-	name string
-	size int64
-	srv  *http.Server
-	t    *testing.T
-}
 
 // TestDownloadToFile tests downloadToFile function, using non-secure protocol(s).
 func TestDownloadToFile(t *testing.T) {
@@ -72,9 +51,8 @@ func TestDownloadToFile(t *testing.T) {
 
 // TestDownloadToFileSecureSystemPool tests downloadToFile function, using secure protocol(s) and certificates from system pool.
 func TestDownloadToFileSecureSystemPool(t *testing.T) {
-	// SystemCertPool does not run on Windows:
-	//https://github.com/golang/go/issues/16736
-	//Fixed in 1.18
+	// SystemCertPool does not work on Windows: https://github.com/golang/go/issues/16736
+	// Fixed in 1.18
 	if runtime.GOOS != "linux" {
 		t.Skip("this test only runs on Linux.")
 	}
@@ -132,8 +110,8 @@ func testDownloadToFile(arts []*Artifact, certFile, certKey string, t *testing.T
 			defer os.RemoveAll(dir)
 
 			// Start Web server
-			srv := host(":43234", art.FileName, int64(art.Size), false, isSecure(art.Link, t), validCert, validKey, t)
-			defer srv.close()
+			srv := Host(":43234", art.FileName, int64(art.Size), false, isSecure(art.Link, t), validCert, validKey, t)
+			defer srv.Close()
 			name := filepath.Join(dir, art.FileName)
 
 			// 1. Resume downlaod of corrupted temporary file.
@@ -208,8 +186,8 @@ func TestDownloadToFileError(t *testing.T) {
 	}
 
 	// Start Web server
-	srv := host(":43234", art.FileName, int64(art.Size), true, isSecure(art.Link, t), extCert, extKey, t)
-	defer srv.close()
+	srv := Host(":43234", art.FileName, int64(art.Size), true, isSecure(art.Link, t), extCert, extKey, t)
+	defer srv.Close()
 	name := filepath.Join(dir, art.FileName)
 
 	// 1. Resume is not supported.
@@ -222,7 +200,7 @@ func TestDownloadToFileError(t *testing.T) {
 	// 2. Try with missing checksum.
 	art.HashValue = ""
 	if err := downloadArtifact(name, art, nil, "", make(chan struct{})); err == nil {
-		t.Fatal("validatted with missing checksum")
+		t.Fatal("validated with missing checksum")
 	}
 
 	// 3. Try with missing link.
@@ -266,21 +244,22 @@ func TestDownloadToFileSecureError(t *testing.T) {
 	defer os.RemoveAll(dir)
 
 	art := &Artifact{
-		FileName: "test.txt", Size: 65536, Link: "https://localhost:43234/test.txt",
+		FileName: "test.txt", Size: 65536,
 		HashType:  "MD5",
 		HashValue: "ab2ce340d36bbaafe17965a3a2c6ed5b",
 	}
 
 	// Start Web servers secure
-	srvSecureInv := host(":43234", art.FileName, int64(art.Size), true, true, expiredCert, expiredKey, t)
-	defer srvSecureInv.close()
-	srvSecureExt := host(":43235", art.FileName, int64(art.Size), true, true, extCert, extKey, t)
-	defer srvSecureExt.close()
-	srvSecureVal := host(":43236", art.FileName, int64(art.Size), true, true, validCert, validKey, t)
-	defer srvSecureVal.close()
+	srvSecureInv := Host(":43234", art.FileName, int64(art.Size), true, true, expiredCert, expiredKey, t)
+	defer srvSecureInv.Close()
+	srvSecureExt := Host(":43235", art.FileName, int64(art.Size), true, true, extCert, extKey, t)
+	defer srvSecureExt.Close()
+	srvSecureVal := Host(":43236", art.FileName, int64(art.Size), true, true, validCert, validKey, t)
+	defer srvSecureVal.Close()
 	name := filepath.Join(dir, art.FileName)
 
 	// 1. Try download with expired certificate
+	art.Link = "https://localhost:43234/test.txt"
 	if err := downloadArtifact(name, art, nil, "", make(chan struct{})); err == nil {
 		t.Fatalf("validated with expired certificate: %v", err)
 	}
@@ -288,10 +267,10 @@ func TestDownloadToFileSecureError(t *testing.T) {
 		t.Fatalf("validated with expired certificate: %v", err)
 	}
 
-	// 2. Try download with unauthorized certificate
+	// 2. Try download with untrusted certificate
 	art.Link = "https://localhost:43235/test.txt"
 	if err := downloadArtifact(name, art, nil, "", make(chan struct{})); err == nil {
-		t.Fatalf("validated with unauthorized certificate: %v", err)
+		t.Fatalf("validated with untrusted certificate: %v", err)
 	}
 
 	// 3. Try download with unknown certificate
@@ -306,121 +285,4 @@ func check(name string, expected int, t *testing.T) {
 	if stat, err := os.Stat(name); os.IsNotExist(err) || stat.Size() != int64(expected) {
 		t.Fatalf("corrupted download artifact: %v != %v", stat.Size(), expected)
 	}
-}
-
-// host create and start a HTTP server.
-func host(addr string, name string, size int64, simple bool, secure bool, cert string, key string, t *testing.T) *web {
-	// Create new HTTP server.
-	w := &web{name: name, size: size, srv: &http.Server{Addr: addr}, t: t}
-
-	// Add HTTP alias only once.
-	if simple {
-		testOnceSimple.Do(func() { http.HandleFunc(fmt.Sprintf("/%s", name), w.handlerSimple) })
-	} else {
-		testOnceRange.Do(func() { http.HandleFunc(fmt.Sprintf("/%s", name), w.handlerRange) })
-	}
-
-	// Start HTTP server in separate goroute.
-	if secure {
-		go func() {
-			if err := w.srv.ListenAndServeTLS(cert, key); err != nil && err != http.ErrServerClosed {
-				t.Errorf("The secure server is not started: %v", err)
-			}
-		}()
-	} else {
-		go func() {
-			if err := w.srv.ListenAndServe(); err != nil && err != http.ErrServerClosed {
-				t.Errorf("Failed to start web server: %v", err)
-			}
-		}()
-	}
-
-	time.Sleep(1 * time.Second)
-
-	return w
-}
-
-// close the http server.
-func (w *web) close() {
-	if err := w.srv.Shutdown(context.Background()); err != nil {
-		w.t.Errorf("failed shutdown web server: %v", err)
-	}
-}
-
-// handlerRange handles incoming HTTP requests with range header support.
-func (w *web) handlerRange(writer http.ResponseWriter, request *http.Request) {
-	// Set default headers for txt file.
-	writer.Header().Set("Content-Type", "text/plain")
-	writer.Header().Set("Content-Disposition", fmt.Sprintf(`attachment; filename="%s"`, w.name))
-	writer.Header().Set("Accept-Ranges", "bytes")
-
-	// Parse range header -> 'Range: bytes=<begin>-<end>'
-	r := request.Header.Get("range")
-	if r == "" { // No range is defined
-		writer.Header().Set("Content-Length", strconv.Itoa(int(w.size)))
-		write(writer, w.size)
-		return
-	}
-
-	// Client requests a part of the file
-	var err error
-	begin, end := int64(0), int64(w.size-1)
-	r = r[6:] // Strip the "bytes=", left over is now "begin-end"
-	pos := strings.Index(r, "-")
-	if pos < 0 {
-		http.Error(writer, "invalid values for header 'Range'", http.StatusBadRequest)
-		return
-	}
-
-	// Parse begin part of the range
-	sb := r[:pos]
-	if len(sb) > 0 {
-		begin, err = strconv.ParseInt(sb, 10, 64)
-		if err != nil || begin > w.size {
-			http.Error(writer, "invalid values for header 'Range' begin", http.StatusBadRequest)
-		}
-	}
-
-	// Parse end part of the range
-	se := r[pos+1:]
-	if len(se) > 0 {
-		end, err = strconv.ParseInt(se, 10, 64)
-		if err != nil || end > w.size || begin >= end {
-			http.Error(writer, "invalid values for header 'Range' end", http.StatusBadRequest)
-		}
-	}
-
-	writer.Header().Set("Content-Length", strconv.FormatInt(end-begin+1, 10))
-	writer.Header().Set("Content-Range", fmt.Sprintf("bytes %d-%d/%d", begin, end, w.size))
-	writer.WriteHeader(http.StatusPartialContent)
-
-	// Send the (end-begin) amount of bytes to the client
-	write(writer, end-begin+1)
-}
-
-// handlerSimple handles incoming HTTP requests without range header support.
-func (w *web) handlerSimple(writer http.ResponseWriter, request *http.Request) {
-	writer.Header().Set("Content-Type", "text/plain")
-	writer.Header().Set("Content-Disposition", fmt.Sprintf(`attachment; filename="%s"`, w.name))
-	writer.Header().Set("Content-Length", strconv.Itoa(int(w.size)))
-	write(writer, w.size)
-}
-
-// write file content.
-func write(writer http.ResponseWriter, size int64) {
-	b := []byte("11111111111111111111111111111111111111111111111111")
-	s := int64(len(b))
-	for i := int64(s); i < size; i += s {
-		writer.Write(b)
-	}
-	remainder := size % s
-	writer.Write(b[:remainder])
-}
-
-func isSecure(link string, t *testing.T) bool {
-	u, err := url.Parse(link)
-	if err != nil {
-		t.Fatalf("cannot parse url %v", link)
-	}
-	return u.Scheme == "https"
 }
