+++ internal/storage/testdata/valid_key.pem
@@ -1,27 +1,52 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEowIBAAKCAQEAoVPo6FuN42eI66uHe4nKOeVdtnEMeJjnf7sJZFTODKYXHcih
-j25wJ0lIdJLE/aQeIU+W+WxjShrkXjjrp5X0VWeYV0J2K8wlG2PBM+Acu+2h7Fw5
-RNAeZwjPnr2pYHN2QjJJRCx8sbQoz+x3dQHWKDcHxNUp1AisL89QRfO1iHZ4LECP
-pcXPR5NNoBP9Jg1XyqKjY7E/YeTjZm5SoDM3iiVRcgQlmVt8vBIir2Fxcjr0PNRW
-A9PwrfMLMReVP98wQDjxKn7Ia+NoYfTZOFE7iX2Z3S5OEWW08K4sevGo+JgL43RN
-O+cykUxLGU+HV2U5VPilNZtMAMaonD2KY0HupQIDAQABAoIBAFmQXB2jtoIpWLp9
-4BmVxc+epwhCesn5GYywKeaprjsXen3uhMM8bMdix/bwBK1MD+4uYYLXyce6Zk5p
-C2tetOc1PD41h6WJEJDEMN677TfoFZ1DFgg+AqLxIVLfuMjzURQvFr3idNezTXUh
-srAi9OIeqxF7xl0Q84cXcVlkk8POH90D/AHb7US/1V3h4oQoCtFe/QdBEXIH+11X
-F6Y97popiq9cm7a7KRcB3KMeTfPpS4X069W8+psxlKIR0oVoxnT1BE6HHxHByBVX
-VvV51vQkTXa2TJpVl52wwXSArbmZg2RCKoCaZV5YR1Kwx+DnHDfcvWJPaVeTf2p5
-/iIxuAECgYEA0ryKs90fzbs/NLSrkkakrPw1HGDCDCbrEZw/XsAkAn7HGGUrK81b
-CaDImDprQ0A7iWjc6n09bl41YCi5uypaa3FzTqRI6zdFJOyGytdEjPSnrgPkMB5Q
-B6K2dU72hyLIuL4ZxKxnmAeAXZ4oCbLBhK0Kqn/5h3FyCwmRHELC3DECgYEAw/qa
-cFgg1WnJq0kiKqlMMUUD1DIBQ4g/+Sd+DyM6g+dmEbK8FAD6JYabzxgCN/lSDYHI
-PNgIGMT12GAhl0Zo/cmgFOCkUakETTPqB4jImRfomQb4Izx4KwepJI9CQ9+qa7QP
-51Sy5ZDK/LbgJ7ws2DzOTux6u/RyZjMHjpNwQLUCgYEAkuRdgnZWHt+wdRqipXmc
-e//pPXisii/ulv3kpjUkZKDeQsDyULv6A0IR0RqXfe8eEPAJXWIjGfWNGxQIMtCY
-6l/QUjbtpReHmO9OHcHjxKUc5I1N18SpYEkpCaga66NxamUILTFHnI9OCPOfUPNL
-NSIxVxuRYbnQFqV6gsp+bWECgYBI7cTy8ZimUkud1gVE4uNbolcea81UmCtWSH/8
-mht4Ym/1PhIHT6XGuSTaGBEbS0hhhl5BZhL9RK5H6t4VMTp41sMfgckkKkVThXId
-g6VCc0c3SwSj+CivupYS6My7/Uup8MPKThaUnyEqkdMwdjIp9dbAP/gcYYcAyRSS
-1gL70QKBgF83mv0E6rmQB1nwE2BD8ZfoPAwtCvZlYNlhCJGLKwSmI9nUITJZl0jH
-IH3XPH/nUKQKJMd8UWMDJwCWH6NrKJ/lePOkHlGx5hQgfirTkofJuLJc25Wpt7At
-/zGhDL2xSfH/RKMijU3+SVe5usDQkzgWz/8yepkyJgR855a8ST+k
------END RSA PRIVATE KEY-----
+-----BEGIN PRIVATE KEY-----
+MIIJQgIBADANBgkqhkiG9w0BAQEFAASCCSwwggkoAgEAAoICAQDFpDqcG3BOqY9k
+Nv4VDCrw2shgrUw/DQT0KYUhi3mWItoGhfPF3Ufrmj5+aGGjhHNZt3GYFOpxdwnq
+70DxRMCXY//UI6nphKyxU3NJqog6kaBJ4cIR7rfVB+frCMNA+HZI4lrQJDgHzjyj
+H1zIh+wYnM0srPEFw2b3gpD+fKz/Bp37oEsSK8bncmFPpJ/il48xlYSsPTPOUmuk
+cL7+CnXLe5t9ryb5hbDPz2WA1Mq3oo5fHEK7dAMJQkRk5nNQSaw9U+m0lXo2YTcz
+CgHhsDevZ813LG/Mg1xdI9/L+LVv9mHD27NYpFoclNpWn+nK0fkXPDpFhDathX6A
+8AVZuH+LobRMrGpsCmubiEoYzh+3F+ilVHspnLWka/MPWMJQ69WDWSY2KDagb9zR
+xfrhjZ7XfNxO4qf9sU7g217LZwWBrUngbSo4E6STBcVfvlQLn1A/9HepbGwX73dX
+yvPWqMugggePxfliOXbASxtNHQL2XYFbDM2IdcgytlvW0bA57g12rqr2X22s9aPS
+JWGfrOyoZHRa1jf/HXRGFbtulWl+K6T4a+QN5L7GeHxzBxsoT5owdtYmEIG86RFd
+1peWA/393Ui/cVxmD1iKt8jwlJh/Q3AnznHN9nF2zdwEbGSEnN88eTYari2h9VFM
+bKtfhRYNpasyldNuEFQ9zvamy+UgqQIDAQABAoICAFIIopMdCHZ+BphwAvI1hLVH
++he2loktHDtD6JBN1f+tI5XFj7M2750o/nt1U+GNHlvEuo+HTvAdmq469FewU8gC
+p7Mjwmi8uvk3TAjtkXHN9Bv2apkK3p67k1WI45JaDq6d1VfDUMZcoeqLH/XUBPkK
+TFcYyUm9G0XsmcYndZ683mIM3FwEc6CpZx6pk/t3MOZByzm6hYBD0Y4FhiDnspDt
+E1MNmj3DceoviU/6AtJJ1JJBpqkrs/7p1H2U8ZQHoEgsRjrAOGPN4h17L+C/Rb22
+dh/a/+SBOZsnQsLkzlpUh6vFyyElxolKd32xDlcy5BLy2rTz9cX8PM/D0FS3MajG
+XZWQUKVUpAF2Sy2zt/WCZiekyrOdx/9bQoEUaA53iATquFohz/WR+oefMvx4MjDc
+z4RnvXD7MX9wnmoqX8+eQc80VKBVE1ygdERWHese5Sk43ECkW9ju/jWfLMSejRo7
+NaL8Wy/qq72lputqJ1ivRigSayCPvi81r0oSA3VnOJXUxxdnm722doD+VLL+Argq
+uKH7F95sxUNkQ81MsyO08X8Bg5lHa1bFtoJUKXH+t0X7UXkvutNfQS9F2y771U3O
+UIanUDol+PD4xJkpFtHhYiRD9+xh2glz1lfBATqws2OgAqBJBykh0w1w+LxbQn+f
+4M5jXreCPILqXia1tRABAoIBAQDroPURMI0tOeg7Ckq8rD4MKNi5br9xJJe1D8Rg
+TNs/uUF7ZReKO6uMb6rcv1Q6aAMJqir6sOhkIEt4Z5C5UlKYmYunO4W8rC8z29Fk
+fDU1TsleFyVHWOl2YPgJqCvnlYFZILfIOdANMSDWWIpVEFG+2lfktLbMh10ZT27R
+zaMOGGx8BSvQMZx2nlA1OWAVozEa3ebLE+Tp0XGN57SSfhPNUiSITvFnqNaGTbIl
+dzuiixHhzhCceqFcKSJEjPHtcCmbeFz0C+nIKA+0fS3fck5yffL6xW2qn4Elb9uA
+3eeVj0OVnMW14fOJOOkuOJPgiJtv6jECwmdfAEAEks7AFh0BAoIBAQDWuoVrn3T/
+BhapOh7mnXQXoQ/xQCv6LlEaLQ27yMu/9pI2BK7eVRU04z7zEiV9sIYfIfEX40aI
+xLP4soCOxn7ZModsGI0EqfX5xOu0hpu84IEqTiGzAhQwesjOp9dLWYmiOwO3Ec1o
+Vi7xah4KHowWsS60pkIYsTGUZ4bMlJO5oqoKn4DAVLCwhY9LQmFB0X3v2uJOZNrl
+lWPAskPSiuRHWWWoMWJT+Olno7uRNyZdR2Qf2Ztvz73HGZkk9oMe9v3qAW0Wf+J6
+/xnksfcq+dr8PKHmb+9Yumtt4za5CRw14Kx+4RkUwoqrcukIEc+wiMBH21X4qHX6
+o+Fq1rli3PupAoIBAAYl5v8R0z0w25xewmB15hCPKEh+WURsfgDdbnRMzRO2k7zS
+gHC86db6Ou0TmPEFMb8Yz5JG92lVZmK34RDU0w7kQr6UUQMBQyIJREdo7kJ70clA
+tVpNOZo7IX2vA6ZBQmS+88ynpYQSmha9rK5XhKqEK/LtvC9OAOHeLrgOs4h0nrAF
+uoTuJnT54IHjA1zNK0TWMWVvGXP4YQwg1ZdU7tNWp1lfJfFceGx8rn9HwyiUlFzd
+8ZUDph4vfJNpUg6t9If39gTwIqN0WCSPd5jy9JUMMEDW0eu3ybjeWE3efGJsoIjQ
+1F/Qi3gzIfJwy3r6VYPwCmNWCzhxWzpDUl20tgECggEATo5eUUk83C+2CqMguu3P
+kvVxh6kabaSUNydUZubG46upQenkr6Ig6ABa/0IcRSdXafxY34xeOZr83HcxYQpt
+TbiWoo04ygQD4SSRtrGoOClKF9gNrBeOm03QOQpeUUqod5lveoIgy5L8CYXxB3K1
+fXfExoIsnCv3H3vgbOfOQ5Om8cv4pmRVX7789WOLgTDLCfzh3hw3mNZWcCDHfzcg
+3Jlv/2BRwLbHbQsRi8bJWkbL8nsByp4Np9FNnaAxyBpEOl/bMOgIS0cYifoxmp9d
+5fmooU1CGwvYUZNWactTbsKLWL+xFU82vBKF/lXFzae58I/lm9rFjrR1u0hi+jAu
+QQKCAQEAo3gRG3Hsbs856ZCGBGrjfHDI4vHenx6WgFYM5Ey5WptU1xrxlJQzkkVB
+rSFgt915blLA3blBWNDnEy3pl6zPQfOr/HNsNQwI384ffzMBeHD55fJBvQ39njSJ
+zG+8MX9a5bbQ3BbpZu+yfo/h+mb30gcoCqGdq2PgAA+2pvbbeC1xaiNIBsdzn4Kj
+Jn5GehPXS5kQsahdJJOLC4hPtch+iWBsVMpFm2NLezo/FjKksv1z+TC5K8Euw4yS
+KuzDzkNvH4YKkl/Z7OS/zKNLNVt/owvHAHLxF/bUjs3dwIXCKf7vJkpKHXPfek8N
+Cjc4qWWlSWxjX5Wd4yH/qTLN2vwT/Q==
+-----END PRIVATE KEY-----
